% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(lang = "en", name = "")
}
\arguments{
\item{\describe{}{\item{lang}{ a \code{character} vector of length 1 thaht correspond to one of the code in the dataset language (see ?language) }
\item{name}{ a \code{character} vector of lenth 1 that specifies the name of the person to whom the message is}
}}
}
\value{
\code{"Hello, world!"}
}
\description{
\code{hello} says \emph{"Hello, world!"}
}
