% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\docType{data}
\name{language}
\alias{language}
\title{The \code{language} dataset}
\format{
this is a \code{data.frame} with 5 rows and 2 column
\describe{
\item{}{\emph{\strong{code}}: Language code: "en", "fr", "it", "es", "de" for corresponding language}
\item{}{\emph{\strong{hello}}: word "hello" in the language corresponding to the code}
}
}
\usage{
language
}
\description{
\code{language} data provides the word "hello" in different languages
}
\author{
Coco
}
\keyword{datasets}
